Move all components to component folder within src
